smart-dashboard-poc/requirements.txt
fastapi[all]==0.111.*
uvicorn==0.29.*
affinda==3.*
openai>=1.0.0
weaviate-client==4.*
playwright==1.*
celery==5.*
redis==5.*
python-dotenv==1.*
pytest==8.*
pyresparser==1.0.6
beautifulsoup4==4.12.*
requests==2.31.*
google-cloud-documentai>=2.20.1
PyPDF2>=3.0.0
python-docx>=1.0.0
